fix マイグレーションできない問題の解消

+++ database/migrations/2020_09_24_082954_create_greets_table.php
@@ -14,8 +14,8 @@ class CreateGreetsTable extends Migration
     public function up()
     {
         Schema::create('greets', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->integer('user_id')->comment('ユーザID');
+            $table->increments('id');
+            $table->unsignedInteger('user_id')->comment('ユーザID');
             $table->string('name')->null()->comment('ユーザ名');
             $table->text('greet')->comment('本文');
             $table->timestamps();
